% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_pseudotime.R
\name{allelicSwitch}
\alias{allelicSwitch}
\title{Performs a number of test to evaluate changes in allelic distribution between the groups.}
\usage{
allelicSwitch(
  a1_counts,
  tot_counts,
  metadata,
  split.var = "group",
  min_counts = 0,
  min_cells = 5,
  estimates,
  estimates_group
)
}
\arguments{
\item{a1_counts}{Reference allele counts}

\item{tot_counts}{Total gene counts counts}

\item{metadata}{Metadata object containing cell level information.
Metadata rownames must match the column names of the count matrices.}

\item{split.var}{Name of the variable which will be used to split the cells.}

\item{min_counts}{Minimum number of mapped reads per cell.
Cells with a number of mapped reads less than min_counts are excluded from the estimation}

\item{min_cells}{Minimum number of cells per gene. Genes with a number of cells less than
min_cells are excluded from the estimation. Default value is 5.}

\item{estimates}{output of bb_init_params()}

\item{estimates_group}{output of estim_bbparams_bygroup()}
}
\description{
Performs a number of test to evaluate changes in allelic distribution between the groups.
}
\examples{
allelicSwitch()
}
