% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_imbalance.R
\name{bb_mean}
\alias{bb_mean}
\title{Likelihood ratio test for mean allelic imbalance (per gene)}
\usage{
bb_mean(
  a1_counts,
  tot_counts,
  estimates,
  glob_params,
  min_cells = 5,
  min_counts = 5,
  batch = NULL,
  metadata = NULL,
  estimates_group = NULL
)
}
\arguments{
\item{a1_counts}{Integer matrix (genes x cells): allele-1 counts.}

\item{tot_counts}{Integer matrix (genes x cells): total counts.
(same dimenstions and rownames as \code{a1_counts}).}

\item{estimates}{Data frame from \code{correct_theta()}}

\item{glob_params}{Named vector/list with elements \code{alpha}, \code{beta}, \code{mu}, \code{theta}
as returned by \code{\link[=glob_disp]{glob_disp()}}.}

\item{min_cells}{Integer >= 1. Minimum number of cells per gene to fit (default 5).
Genes with a number of cells less than min_cells are excluded from the estimation.}

\item{min_counts}{Integer >= 0. Minimum reads per cell to include (default 0).
Cells with a number of mapped reads less than min_counts are excluded from the estimation}

\item{batch}{Optional string: column name in \code{metadata} identifying batches.
(if batch correction is required)}

\item{metadata}{Optional metadata object containing cell level information
(batch identifier must be one of the column in the cell metadata)}

\item{estimates_group}{Optional object containing initial beta-binomial parameter estimates and correction
performed on each batch separately}
}
\description{
Tests H0: mu = mu0 (global/theoretical mean) vs H1: mu = gene-specific mean,
using beta-binomial likelihoods with shrunk dispersion from \code{correct_theta()}.
}
\examples{
bb_mean(a1_counts,
        tot_counts,
        estimates,
        glob_params,
        min_cells = 5,
        min_counts = 5)
}
