% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_variation.R
\name{bb_var}
\alias{bb_var}
\title{Test to evaluate deviation from the expected level of allelic variation for genes with
similar expression. Performs a permutation test between H_0: gene's dispersion is the same
as the common (expected) dispersion. H_1: The stabilized (shrunk) dispersion is not the same as
the common dispersion.}
\usage{
bb_var(
  a1_counts,
  tot_counts,
  estimates,
  estimates_group = NULL,
  min_cells = 5,
  min_counts = 5,
  batch = NULL,
  metadata = NULL,
  n_pmt = 500,
  n_sim = 500
)
}
\arguments{
\item{a1_counts}{Integer matrix (genes x cells): allele-1 counts.}

\item{tot_counts}{Integer matrix (genes x cells): total counts
(same dimenstions and rownames as \code{a1_counts}).}

\item{estimates}{Data frame from \code{correct_theta()}}

\item{estimates_group}{Optional object containing initial beta-binomial parameter estimates and correction
performed on each batch separately}

\item{min_cells}{Integer >= 1. Minimum number of cells per gene to fit (default 5).
Genes with a number of cells less than min_cells are excluded from the estimation.}

\item{min_counts}{Integer >= 0. Minimum reads per cell to include (default 0).
Cells with a number of mapped reads less than min_counts are excluded from the estimation}

\item{batch}{Optional string: column name in \code{metadata} identifying batches.
(if batch correction is required)}

\item{metadata}{Optional metadata object containing cell level information
(batch identifier must be one of the column in the cell metadata)}

\item{n_pmt}{Integer. Number of permutations (<= n_sim.)}

\item{n_sim}{Integer. Number of simulated replicates.}
}
\description{
Test to evaluate deviation from the expected level of allelic variation for genes with
similar expression. Performs a permutation test between H_0: gene's dispersion is the same
as the common (expected) dispersion. H_1: The stabilized (shrunk) dispersion is not the same as
the common dispersion.
}
\examples{
bb_var()
}
