% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_imbalance.R
\name{beta_binom_test}
\alias{beta_binom_test}
\title{Performs likelihood ratio test between H_0: no allelic imbalance
(mean allelic ratio = 0.5, or another theoretical value based on the global parameter estimates)
and H_1: allelic imbalance is present. Likelihoods for H_0 and H_1 are calculated based on
the mu and theta parameters from correct_theta function}
\usage{
beta_binom_test(
  a1_counts,
  tot_counts,
  estimates,
  glob_params,
  min_cells = 5,
  min_counts = 0,
  batch = NULL,
  metadata = NULL
)
}
\arguments{
\item{a1_counts}{Reference allele counts}

\item{tot_counts}{Total gene counts counts}

\item{estimates}{output from correct_theta()}

\item{glob_params}{output from the glob_disp}

\item{min_cells}{Minimum number of cells per gene. Genes with a number of cells less than
min_cells are excluded from the estimation}

\item{min_counts}{Minimum number of mapped reads per cell.
Cells with a number of mapped reads less than min_counts are excluded}
}
\description{
Performs likelihood ratio test between H_0: no allelic imbalance
(mean allelic ratio = 0.5, or another theoretical value based on the global parameter estimates)
and H_1: allelic imbalance is present. Likelihoods for H_0 and H_1 are calculated based on
the mu and theta parameters from correct_theta function
}
\examples{
beta_binom_test()
}
