% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{calc_mad}
\alias{calc_mad}
\title{Calculating MAD^2 on the dispersion trend residuals}
\usage{
calc_mad(estimates)
}
\arguments{
\item{estimates}{Data frame from \code{estim_bbparams()}}
}
\description{
Calculating MAD^2 on the dispersion trend residuals
}
\examples{
calc_mad(estimates)
}
