% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{correct_theta}
\alias{correct_theta}
\title{Fits a local regression model with dispersion as a function of total gene counts.
Predicted values are the expected dispersion for genes with similar expression levels
The initial dispersion estimates are shrunk towards the trend}
\usage{
correct_theta(
  estimates,
  delta_set = 50,
  N_set = 30,
  thetaFilter = NULL,
  shrinkAll = FALSE
)
}
\arguments{
\item{estimates}{Data frame from \code{estim_bbparams()}}

\item{delta_set}{Delta parameter (numeric > 0).}

\item{N_set}{N parameter (numeric > 0).}

\item{thetaFilter}{Optional numeric threshold; genes with bb_theta < thetaFilter are excluded.
Genes with dispersion below the set threshold are excluded from the shrinking procedure.}

\item{shrinkAll}{Logical. If TRUE, apply shrinkage to all genes (default FALSE).}
}
\description{
Fits a local regression model with dispersion as a function of total gene counts.
Predicted values are the expected dispersion for genes with similar expression levels
The initial dispersion estimates are shrunk towards the trend
}
\examples{
correct_theta(estimates,
delta_set = 50,
N_set = 30,
thetaFilter = 1e-3,
shrinkAll = FALSE)
}
