% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_imbalance.R
\name{estim_bbparams}
\alias{estim_bbparams}
\title{Estimate beta-binomial distribution parameters. This is the initial function that estimates
beta-binomial distribution parameters for the allele-specific counts for each gene. The users
may filter out low abundant or low expressed genes through quality control parameters of
minimum number of reads per cell and minimum number of cells per gene.}
\usage{
estim_bbparams(
  a1_counts,
  tot_counts,
  min_counts = 0,
  min_cells = 5,
  cores = NULL
)
}
\arguments{
\item{a1_counts}{Reference allele counts}

\item{tot_counts}{Total gene counts counts}

\item{min_counts}{Minimum number of mapped reads per cell.
Cells with a number of mapped reads less than min_counts are excluded from the estimation}

\item{min_cells}{Minimum number of cells per gene. Genes with a number of cells less than
min_cells are excluded from the estimation. Default value is 5.}

\item{cores}{number of processing cores}
}
\description{
Estimate beta-binomial distribution parameters. This is the initial function that estimates
beta-binomial distribution parameters for the allele-specific counts for each gene. The users
may filter out low abundant or low expressed genes through quality control parameters of
minimum number of reads per cell and minimum number of cells per gene.
}
\examples{
estim_bbparams()
}
