% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_pseudotime.R
\name{estim_bbparams_bygroup}
\alias{estim_bbparams_bygroup}
\title{Estimate beta-binomial distribution parameters. This is the initial function that estimates
beta-binomial distribution parameters for the allele-specific counts for each gene. The users
may filter out low abundant or low expressed genes through quality control parameters of
minimum number of reads per cell and minimum number of cells per gene.}
\usage{
estim_bbparams_bygroup(
  a1_counts,
  tot_counts,
  metadata,
  split.var = "group",
  min_counts = 0,
  min_cells = 5,
  cores = NULL
)
}
\arguments{
\item{a1_counts}{Reference allele counts}

\item{tot_counts}{Total gene counts counts}

\item{metadata}{Metadata object containing cell level information.
Metadata rownames must match the column names of the count matrices.}

\item{split.var}{Name of the variable which will be used to split the cells.}

\item{min_counts}{Minimum number of mapped reads per cell.
Cells with a number of mapped reads less than min_counts are excluded from the estimation}

\item{min_cells}{Minimum number of cells per gene. Genes with a number of cells less than
min_cells are excluded from the estimation. Default value is 5.}

\item{cores}{number of processing cores}
}
\description{
Estimate beta-binomial distribution parameters. This is the initial function that estimates
beta-binomial distribution parameters for the allele-specific counts for each gene. The users
may filter out low abundant or low expressed genes through quality control parameters of
minimum number of reads per cell and minimum number of cells per gene.
}
\examples{
estim_bbparams_bygroup()
}
