% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geBoxplot}
\alias{geBoxplot}
\title{Boxplot of allele-specific expression}
\usage{
geBoxplot(plot_data, allele1 = "allele1", allele2 = "allele2")
}
\arguments{
\item{plot_data}{output of make_plotdf_simul}

\item{allele1}{names of the non-reference strain}
}
\description{
Boxplot of allele-specific expression
}
\examples{
geBoxplot()
}
