% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{glob_disp}
\alias{glob_disp}
\title{Estimates global beta-binomial distribution parameters on all genes together to
evaluate the degree of skewness towards the reference allele. For the estimation,
genes located on the sex chromosomes and imprinted genes are excluded}
\usage{
glob_disp(a1_counts, tot_counts, genes.excl, min_counts = 5)
}
\arguments{
\item{a1_counts}{Reference allele counts}

\item{tot_counts}{Integer matrix (genes x cells): total counts
(same dimenstions and rownames as \code{a1_counts}).}

\item{genes.excl}{Character vector of gene IDs to exclude
(eg. sex chromosome or imprinted genes)}

\item{min_counts}{Integer >= 0. Minimum reads per cell to include (default 0).
Cells with a number of mapped reads less than min_counts are excluded from the estimation}
}
\description{
Estimates global beta-binomial distribution parameters on all genes together to
evaluate the degree of skewness towards the reference allele. For the estimation,
genes located on the sex chromosomes and imprinted genes are excluded
}
\examples{
glob_disp(a1_counts,
          tot_counts,
          genes.excl,
          min_counts = 5)
}
