% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffASE.R
\name{group_mean}
\alias{group_mean}
\title{Performs test to detect changes in allelic ratio distribution across discrete groups.}
\usage{
group_mean(
  a1_counts,
  tot_counts,
  metadata,
  split.var = "group",
  min_counts = 0,
  min_cells = 5,
  estimates,
  estimates_group,
  equalGroups = TRUE
)
}
\arguments{
\item{a1_counts}{Integer matrix (genes x cells): allele-1 counts.}

\item{tot_counts}{Integer matrix (genes x cells): total counts.
(same dimenstions and rownames as \code{a1_counts}).}

\item{metadata}{Metadata object containing cell level information
(group identifier must be one of the column in the cell metadata)}

\item{split.var}{Name of the variable (group identifier) which will be used to split the cells.}

\item{min_counts}{Integer >= 0. Minimum reads per cell to include (default 0).
Cells with a number of mapped reads less than min_counts are excluded from the estimation}

\item{min_cells}{Integer >= 1. Minimum number of cells per gene to fit (default 5).
Genes with a number of cells less than min_cells are excluded from the estimation.}

\item{estimates}{Data frame from \code{correct_theta()}}

\item{estimates_group}{a list where each element is a data frame with
group-level betabinomial estimates and corrected dispersion for each gene}

\item{equalGroups}{Controls for the equal number of cells between the groups (default TRUE)}
}
\description{
Tests H0: all groups share the same mean vs H1: allelic ratio means are group-specific
}
\examples{
group_mean()
}
