% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_plotdf_simul}
\alias{make_plotdf_simul}
\title{posterior counts simulated under the expected dispersion}
\usage{
make_plotdf_simul(
  a1_counts,
  tot_counts,
  gene,
  metadata,
  split.var = "group",
  order.by = "time",
  estimates_group,
  init.seed = 10001101
)
}
\arguments{
\item{a1_counts}{Reference allele counts}

\item{tot_counts}{Total gene counts counts}

\item{gene}{gene symbol or other id, the same style that is used for the row names in the count matrices}

\item{metadata}{Metadata object containing cell level information.
Metadata rownames must match the column names of the count matrices.}

\item{split.var}{Name of the variable which will be used to split the cells.}

\item{order.by}{variable to order allelic ratio by}

\item{estimates_group}{output of estim_bbparams_bygroup()}

\item{init.seed}{set seed value}
}
\description{
posterior counts simulated under the expected dispersion
}
\examples{
make_plotdf_simul()
}
