% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makedf}
\alias{makedf}
\title{preparing data frame for plotting}
\usage{
makedf(
  a1_counts,
  tot_counts,
  gene,
  metadata = NULL,
  order.by = "time",
  split.var = "group"
)
}
\arguments{
\item{a1_counts}{Reference allele counts}

\item{tot_counts}{Total gene counts counts}

\item{gene}{gene symbol or other id, the same style that is used for the row names in the count matrices}
}
\description{
preparing data frame for plotting
}
\examples{
makedf()
}
