% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{plot_disp}
\alias{plot_disp}
\title{Plot to visualize original and shrunk dispersion}
\usage{
plot_disp(param_reestim)
}
\arguments{
\item{param_reestim}{Data frame from \code{correct_theta()}}
}
\description{
Plot to visualize original and shrunk dispersion
}
\examples{
plot_disp()
}
