% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{plot_disp_fit_mean}
\alias{plot_disp_fit_mean}
\title{Plot to visualize allelic ratio modelled as a function of total gene counts}
\usage{
plot_disp_fit_mean(param_reestim, midpoint = 2000)
}
\arguments{
\item{param_reestim}{Data frame from \code{correct_theta()}}
}
\description{
Plot to visualize allelic ratio modelled as a function of total gene counts
}
\examples{
plot_disp_fit_mean()
}
