% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_distr}
\alias{plot_distr}
\title{plotting allelic ratio distribution}
\usage{
plot_distr(plot_data, gene, add.density = TRUE, min_counts = 0)
}
\arguments{
\item{plot_data}{output of makedf}

\item{gene}{gene symbol or other id, must match the gene used to generate data frame for plotting}
}
\description{
plotting allelic ratio distribution
}
\examples{
plot_distr()
}
