% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_exp_disp}
\alias{plot_exp_disp}
\title{plotting mean gene expression vs allelic dispersion}
\usage{
plot_exp_disp(param_estim, gene = gene)
}
\arguments{
\item{param_estim}{output of bb_test_res or correct_theta}

\item{gene}{gene symbol or other id, the same style that is used for the row names in the count matrices}
}
\description{
plotting mean gene expression vs allelic dispersion
}
\examples{
plot_exp_disp()
}
