% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{plot_glob_params}
\alias{plot_glob_params}
\title{Plot histogram of allelic ratio distribution across all genes
with beta function based on alpha and beta global parameters}
\usage{
plot_glob_params(a1_counts, tot_counts, glob_params, min_counts = 5)
}
\arguments{
\item{a1_counts}{Integer matrix (genes x cells): allele-1 counts.}

\item{tot_counts}{Integer matrix (genes x cells): total counts
(same dimenstions and rownames as \code{a1_counts}).}

\item{glob_params}{Named vector/list with elements \code{alpha}, \code{beta}, \code{mu}, \code{theta}
as returned by \code{\link[=glob_disp]{glob_disp()}}.}

\item{min_counts}{Integer >= 0. Minimum reads per cell to include (default 0).
Cells with a number of mapped reads less than min_counts are excluded}
}
\description{
Plot histogram of allelic ratio distribution across all genes
with beta function based on alpha and beta global parameters
}
\examples{
plot_glob_params(a1_counts,
                 tot_counts,
                 glob_params,
                 min_counts = 5)
}
