% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_MA}
\alias{plot_MA}
\title{Plotting distribution of allelic ratio over mean expression}
\usage{
plot_MA(res, fdr_var = FALSE, fdr_cutoff = 0.05, min_logFC = 1, glob_params)
}
\arguments{
\item{res}{output file from beta_binom_test}

\item{fdr_var}{Binary parameter to specify which test fdr's to plot (default: fdr_var = FALSE, allelic imbalance results are used for plotting )}

\item{fdr_cutoff}{numeric value of the fdr cut-off to use (default: fdr_cutoff = 0.05)}

\item{min_logFC}{Log FC value to use}

\item{glob_params}{output from glob_disp()}
}
\description{
Plotting distribution of allelic ratio over mean expression
}
\examples{
plot_MA()
}
