% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_theta_density}
\alias{plot_theta_density}
\title{Density plot comparing allelic ratio distribution under observed and expected dispersion}
\usage{
plot_theta_density(plot_data)
}
\arguments{
\item{plot_data}{output of make_plotdf_simul}
}
\description{
Density plot comparing allelic ratio distribution under observed and expected dispersion
}
\examples{
plot_exp_disp()
}
