% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{estim_bbparams}
\alias{estim_bbparams}
\title{Estimate beta-binomial distribution parameters per gene.
The users can remove low abundant or low expressed genes through quality control parameters
\code{min_counts} and \code{min_cells}.}
\usage{
estim_bbparams(
  a1_counts,
  tot_counts,
  min_counts = 0,
  min_cells = 5,
  cores = NULL
)
}
\arguments{
\item{a1_counts}{Integer matrix (genes x cells): allele-1 counts.}

\item{tot_counts}{Integer matrix (genes x cells): total counts
(same dimenstions and rownames as \code{a1_counts}).}

\item{min_counts}{Integer >= 0. Minimum reads per cell to include (default 0).
Cells with a number of mapped reads less than min_counts are excluded from the estimation}

\item{min_cells}{Integer >= 1. Minimum number of cells per gene to fit (default 5).
Genes with a number of cells less than min_cells are excluded from the estimation.}

\item{cores}{number of processing cores (default 1).}
}
\description{
Estimate beta-binomial distribution parameters per gene.
The users can remove low abundant or low expressed genes through quality control parameters
\code{min_counts} and \code{min_cells}.
}
\examples{
estim_bbparams(a1_counts,
              tot_counts,
              min_counts = 0,
              min_cells = 5,
              cores = NULL)
}
