% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimation.R
\name{estim_delta}
\alias{estim_delta}
\title{Estimating appropriate tuning paramaters delta and the number of degrees of freedom, N.
We assume that dispersion follows Gamma distribution. Appropriate shrinkage is selected based on
the MLE of the the difference between fitted dispersion and shrunk dispersion}
\usage{
estim_delta(estimates, thetaFilter = NULL)
}
\arguments{
\item{estimates}{Data frame from \code{estim_bbparams()}}

\item{thetaFilter}{Optional numeric threshold; genes with bb_theta < thetaFilter are excluded.
Genes with dispersion below the set threshold are excluded from the shrinking procedure.}
}
\description{
Estimating appropriate tuning paramaters delta and the number of degrees of freedom, N.
We assume that dispersion follows Gamma distribution. Appropriate shrinkage is selected based on
the MLE of the the difference between fitted dispersion and shrunk dispersion
}
\examples{
estim_delta()
}
