% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_imbalance.R
\name{glob_disp}
\alias{glob_disp}
\title{Estimates global beta-binomial distribution parameters on all genes together to
evaluate the degree of skewness towards the reference allele. For the estimation,
genes located on the sex chromosomes and imprinted genes are excluded}
\usage{
glob_disp(a1_counts, tot_counts, genes.excl, min_counts = 5)
}
\arguments{
\item{a1_counts}{Reference allele counts}

\item{tot_counts}{Total gene counts counts}

\item{min_counts}{Minimum number of mapped reads per cell.
Cells with a number of mapped reads less than min_counts are excluded}

\item{genesXY}{list of sex chromosomes genes}

\item{genesIMPR}{imprinted genes}
}
\description{
Estimates global beta-binomial distribution parameters on all genes together to
evaluate the degree of skewness towards the reference allele. For the estimation,
genes located on the sex chromosomes and imprinted genes are excluded
}
\examples{
glob_disp()
}
