% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_imbalance.R
\name{plot_disp}
\alias{plot_disp}
\title{Plot to visualize original and shrunk dispersion}
\usage{
plot_disp(param_reestim)
}
\arguments{
\item{param_reestim}{Output of correct_theta}
}
\description{
Plot to visualize original and shrunk dispersion
}
\examples{
plot_disp()
}
