% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_imbalance.R
\name{plot_glob_params}
\alias{plot_glob_params}
\title{Plot histogram of allelic ratio distribution across all genes
with beta function based on alpha and beta global parameters}
\usage{
plot_glob_params(a1_counts, tot_counts, glob_params, min_counts = 5)
}
\arguments{
\item{a1_counts}{Reference allele counts}

\item{tot_counts}{Total gene counts counts}

\item{glob_params}{output from the glob_disp}

\item{min_counts}{Minimum number of mapped reads per cell.
Cells with a number of mapped reads less than min_counts are excluded}
}
\description{
Plot histogram of allelic ratio distribution across all genes
with beta function based on alpha and beta global parameters
}
\examples{
plot_glob_params()
}
